# Conflicts:
#	README.md
#	scripts/patch_jpegdec.py
#	src/activities/home/HomeActivity.cpp
#	src/activities/home/RecentBooksActivity.cpp
#	src/activities/settings/KOReaderAuthActivity.cpp
#	src/activities/settings/OpdsServerListActivity.cpp
#	src/activities/settings/OtaUpdateActivity.cpp
#	src/main.cpp

+++ lib/hal/HalPowerManager.h
@@ -19,18 +19,18 @@ class HalPowerManager {
   bool isLowPower = false;
 
   // I2C fuel gauge configuration for X3 battery monitoring
-  bool _batteryUseI2C = false;                   // True if using I2C fuel gauge (X3), false for ADC (X4)
-  mutable int _batteryCachedPercent = 0;         // Last read battery percentage (0-100)
-  mutable unsigned long _batteryLastPollMs = 0;  // Timestamp of last battery read in milliseconds
+  bool _batteryUseI2C = false;              // True if using I2C fuel gauge (X3), false for ADC (X4)
+  mutable int _batteryCachedPercent = 0;    // Last read battery percentage (0-100)
+  mutable uint32_t _batteryLastPollMs = 0;  // Timestamp of last battery read in milliseconds
 
   enum LockMode { None, NormalSpeed };
   LockMode currentLockMode = None;
   SemaphoreHandle_t modeMutex = nullptr;  // Protect access to currentLockMode
 
  public:
-  static constexpr int LOW_POWER_FREQ = 10;                    // MHz
-  static constexpr unsigned long IDLE_POWER_SAVING_MS = 3000;  // ms
-  static constexpr unsigned long BATTERY_POLL_MS = 1500;       // ms
+  static constexpr int LOW_POWER_FREQ = 10;               // MHz
+  static constexpr uint32_t IDLE_POWER_SAVING_MS = 3000;  // ms
+  static constexpr uint32_t BATTERY_POLL_MS = 1500;       // ms
 
   void begin();
